feat(cli): add `defineCodemod` to ease creation
Rearranges a bunch of code to make this work and creates the `@codemod/utils` package which is passed to codemods defined using `defineCodemod`.
