+++ packages/cli/__tests__/fixtures/plugin/append-options-string.ts
@@ -1,14 +1,11 @@
-import { PluginObj, NodePath } from '@babel/core'
-import * as t from '@babel/types'
+import { defineCodemod } from '../../../src'
 
-export default function <T>(_: unknown, options: T): PluginObj {
-  return {
-    visitor: {
-      Program(path: NodePath<t.Program>): void {
-        path.node.body.push(
-          t.expressionStatement(t.stringLiteral(JSON.stringify(options)))
-        )
-      },
+export default defineCodemod(({ t }, options) => ({
+  visitor: {
+    Program(path) {
+      path.node.body.push(
+        t.expressionStatement(t.stringLiteral(JSON.stringify(options)))
+      )
     },
-  }
-}
+  },
+}))
