+++ packages/cli/__tests__/fixtures/plugin/increment-typescript.ts
@@ -1,18 +1,14 @@
-import { NodePath } from '@babel/traverse'
-import { PluginObj } from '@babel/core'
-import * as t from '@babel/types'
+import { defineCodemod } from '../../../src'
 
 enum IncrementValues {
   One = 1,
   Two = 2,
 }
 
-export default function (): PluginObj {
-  return {
-    visitor: {
-      NumericLiteral(path: NodePath<t.NumericLiteral>) {
-        path.node.value += IncrementValues.One
-      },
+export default defineCodemod(() => ({
+  visitor: {
+    NumericLiteral(path) {
+      path.node.value += IncrementValues.One
     },
-  }
-}
+  },
+}))
