+++ packages/cli/__tests__/fixtures/prettier/with-config/.prettierrc
@@ -1,4 +1,4 @@
 {
-    "singleQuote": true,
-    "semi": true
+  "singleQuote": true,
+  "semi": true
 }
