+++ packages/cli/__tests__/fixtures/prettier/with-config/index.js
@@ -1,1 +1,1 @@
-var a = "";
+var a = '';
