+++ packages/core/tsconfig.build.json
@@ -9,7 +9,5 @@
     "rootDir": "src",
     "outDir": "build"
   },
-  "references": [
-    { "path": "../parser/tsconfig.build.json" }
-  ]
+  "references": [{ "path": "../parser/tsconfig.build.json" }]
 }
