+++ packages/core/tsconfig.json
@@ -9,7 +9,5 @@
     "strict": true
   },
   "exclude": ["tmp", "build"],
-  "references": [
-    { "path": "../parser/tsconfig.build.json" }
-  ]
+  "references": [{ "path": "../parser/tsconfig.build.json" }]
 }
