+++ packages/matchers/src/matchers/anyExpression.ts
@@ -1,5 +1,4 @@
-import * as t from '@babel/types'
-import { isNode } from '../NodeTypes'
+import { isNode, t } from '@codemod/utils'
 import { Matcher } from './Matcher'
 
 export class AnyExpressionMatcher extends Matcher<t.Expression> {
