+++ packages/matchers/src/matchers/function.ts
@@ -1,5 +1,4 @@
-import * as t from '@babel/types'
-import { isNode } from '../NodeTypes'
+import { isNode, t } from '@codemod/utils'
 import { Matcher } from './Matcher'
 import { tupleOf } from './tupleOf'
