+++ packages/utils/.eslintignore
@@ -1,0 +1,3 @@
+build
+coverage
+src/matchers.ts
