+++ packages/utils/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@codemod/matchers",
-  "version": "1.5.0",
-  "description": "Matchers for JavaScript & TypeScript codemods.",
-  "repository": "https://github.com/codemod-js/codemod",
+  "name": "@codemod/utils",
+  "version": "1.0.0",
+  "description": "codemod rewrites JavaScript and TypeScript",
+  "repository": "https://github.com/codemod-js/codemod.git",
   "license": "Apache-2.0",
   "author": "Brian Donovan",
   "main": "build/index.js",
@@ -21,28 +21,16 @@
     "test:watch": "jest --watch"
   },
   "dependencies": {
-    "@babel/types": "^7.20.7"
+    "@babel/core": "^7.20.12",
+    "@babel/types": "^7.20.7",
+    "@codemod/matchers": "^1.5.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.12",
-    "@babel/generator": "^7.20.14",
-    "@babel/traverse": "^7.20.13",
-    "@codemod/core": "^2.1.0",
-    "@codemod/parser": "^1.3.0",
-    "@codemod/utils": "^1.0.0",
     "@types/babel__core": "^7.20.0",
     "@types/babel__generator": "^7.6.4",
-    "@types/babel__template": "^7.4.1",
     "@types/babel__traverse": "^7.18.3",
-    "@types/dedent": "^0.7.0",
-    "@types/jest": "^25.1.0",
-    "@types/prettier": "^2.0.0",
-    "dedent": "^0.7.0",
+    "@types/jest": "^29.4.0",
     "is-ci-cli": "^2.2.0",
-    "jest": "^27.3.1",
-    "typescript": "^4.9.5"
-  },
-  "publishConfig": {
-    "access": "public"
+    "jest": "^27.3.1"
   }
 }
