+++ packages/utils/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "compilerOptions": {
+    "target": "es2015",
+    "module": "commonjs",
+    "composite": true,
+    "noEmit": true,
+    "strict": true,
+    "esModuleInterop": true
+  },
+  "exclude": ["build"],
+  "references": [
+    { "path": "../core/tsconfig.build.json" },
+    { "path": "../parser/tsconfig.build.json" }
+  ]
+}
